reorganized scripts
