+++ scripts/tools/clear_world.py
