+++ scripts/tools/collision_matrix_tool.py
