+++ scripts/tools/print_joint_state.py
